Added toolBar, garden

+++ Assets/Resources/Scripts/Items/Item.cs
@@ -7,7 +7,7 @@ public abstract class Item : MonoBehaviour
 {
     [SerializeField] public string name;
     [SerializeField] public Sprite image;
-    [SerializeField] public float price;
+    [SerializeField] public int price;
 
     public abstract void Use();
 }
